engine: extract esm/tours/tour_stop.js
